chore(cli): Separate shiny inspect (reactlog) into dedicated PR and streamline developer CLI tools
