+++ shiny/_main/__init__.py
@@ -5,7 +5,6 @@ import click
 from .. import __version__
 from ._create import create
 from ._generate_test import add
-from ._inspect import inspect
 from ._run import run
 from ._run import run_app as run_app  # noqa: F401
 from ._simulate import simulate
@@ -25,7 +24,6 @@ main.add_command(add)
 main.add_command(create)
 main.add_command(validate)
 main.add_command(simulate)
-main.add_command(inspect)
 main.add_command(skills)
 
 main.add_command(static)
